requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.8.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
urllib3>=2.0.0
python-dateutil>=2.8.2
tld>=0.13.0
dnspython>=2.4.0
certifi>=2023.7.22
cryptography>=41.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
rich==13.7.0
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
gunicorn==21.2.0
tabulate>=0.9.0